edit done

+++ Enoca.API/Controllers/CompaniesController.cs
@@ -24,33 +24,47 @@ namespace Enoca.API.Controllers
             _mapper = mapper;
             _companyService = companyService;
         }
-
+        /// <summary>
+        /// Mevcut tüm şirketleri getirmek için HTTP GET isteği.
+        /// </summary>
+        /// <returns></returns>
         [HttpGet]
-        public async Task<IActionResult> GetAllAsync()
-        {
+        public async Task<IActionResult> GetAllCompaniesAsync()
+        {            
             var companies = await _companyService.GetAllAsync();
-            return Ok(companies);
+            if (companies == null )
+            {
+                return NoContent(); // Boş şirket listesi durumunda HTTP 204 Yanıtı döndürür.
+            }
+            return StatusCode(200,companies);
         }
-        [HttpGet("{id}")]
+
+        /// <summary>
+        /// Belirtilen Id değeri ile eşleşen şirketi getirmek için HTTP GET isteği.
+        /// </summary>
+        /// <param name="id">Şirketin Id'si</param>
+        /// <returns>HTTP yanıtı ve istenen şirket verileri</returns>
+        [HttpGet("{id}")]        
         public async Task<IActionResult> GetByIdCompanyAsync(int id)
         {
             var company = await _companyService.GetByIdAsync(id);
             return Ok(company);
         }
-        [HttpPost]
-        public async Task<IActionResult> AddAsync([FromBody] CreateCompanyDto createCompanyDto)
-        {
-            if (createCompanyDto == null)
-            {
-                return BadRequest("Company cannot be null");
-            }
 
-            if (!ModelState.IsValid)
+        /// <summary>
+        /// Yeni şirket ekler.
+        /// </summary>
+        /// <param name="createCompanyDto"></param>
+        /// <returns>Status code ve eklenen şirketin bilgileri</returns>
+        [HttpPost]
+        public async Task<IActionResult> AddCompanyAsync([FromBody] CreateCompanyDto createCompanyDto)
+        {
+            // Gelen istek için gerekli kontroller yapılır ve hatalı ise BadRequest döndürülür.
+            if (createCompanyDto == null && !ModelState.IsValid)
             {
                 return BadRequest(ModelState);
             }
 
-            
 
             TimeSpan startTime;
             TimeSpan finishTime;
@@ -70,43 +84,79 @@ namespace Enoca.API.Controllers
             company.OrderStartTime = startTime;
             company.OrderFinishTime = finishTime;
 
+            // Yeni şirket oluşturulur ve veritabanına ekler.
             await _companyService.AddAsync(company);
 
             var companyDto = _mapper.Map<CompanyDto>(company);
 
-            return Ok(companyDto);
+            return StatusCode(201, companyDto); 
         }
 
-        [HttpPut]
-        public async Task<IActionResult> Update(CompanyDto companyDto)
+
+        /// <summary>
+        /// Mevcut bir şirketi günceller.
+        /// </summary>
+        /// <param name="companyDto">Güncellenecek şirket verileri</param>
+        /// <returns>Status code ve status açıklaması</returns>
+        [HttpPut("{id}")]
+        public async Task<IActionResult> UpdateCompanyAsync(CompanyDto companyDto)
         {
+            if(companyDto == null)
+                return NotFound();
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+                
             await _companyService.UpdateAsync(_mapper.Map<Company>(companyDto)); 
-            return Ok();
+            return StatusCode(200,"Başarılı bir şekilde güncellendi.");
         }
 
+
+        /// <summary>
+        /// Mevcut bir şirketi siler.
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns>Status code ve açıklaması</returns>
         [HttpDelete("{id}")]
-        public async Task<IActionResult> Remove(int id)
+        public async Task<IActionResult> RemoveCompanyAsync(int id)
         {
             var product = await _companyService.GetByIdAsync(id);
+            if(product == null)
+                return NotFound(ModelState);
             await _companyService.RemoveAsync(product);
-            return Ok(true);
+            return StatusCode(200, "Şirket başarıı bir şekilde silindi.");
         }
 
+
+        /// <summary>
+        /// Şirketin Onay durumunu günceller.
+        /// </summary>
+        /// <param name="id"></param>
+        /// <param name="companyStatusUpdateDto"></param>
+        /// <returns>Status code ve açıklaması</returns>
         [HttpPut("{id}")]
         public async Task<IActionResult> UpdateIsStatusAsync(int id, [FromBody] CompanyStatusUpdateDto companyStatusUpdateDto)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             var company = await _companyService.GetByIdAsync(id);
             if (company == null)
             {
                 return NotFound();
-            }
-           
+            }           
             company.IsStatus = companyStatusUpdateDto.IsStatus;
             await _companyService.UpdateAsync(company);
-            return Ok();
-
+            return StatusCode(200, "Onay durumu güncellendi!");
         }
 
+
+        /// <summary>
+        /// Belirtilen ID'li şirketin onay durumunu getirir.
+        /// </summary>
+        /// <param name="id"></param>
+        /// <param name="companyStatusUpdateDto"></param>
+        /// <returns>boolean türünde onay durumu</returns>
         [HttpGet("{id}")]
         public async Task<IActionResult> GetCompanyStatusAsync(int id, CompanyStatusUpdateDto companyStatusUpdateDto)
         {
@@ -115,12 +165,17 @@ namespace Enoca.API.Controllers
             {
                 return NotFound();
             }
-            bool result =  company.IsStatus;
-            return Ok(result);
+            bool status =  company.IsStatus;
+            return Ok(status);
 
         }
 
-
+        /// <summary>
+        /// Belirtilen ID'li şirketin şipariş başlangıç ve bitiş sürelerini günceller.
+        /// </summary>
+        /// <param name="id"></param>
+        /// <param name="orderTimeUpdateDto"></param>
+        /// <returns></returns>
         [HttpPut("{id}")]
         public async Task<IActionResult> UpdateOrderTimeAsync(int id, [FromBody] OrderTimeUpdateDto orderTimeUpdateDto)
         {
@@ -145,7 +200,7 @@ namespace Enoca.API.Controllers
             company.OrderFinishTime = finishTime;
 
             await _companyService.UpdateAsync(company);
-            return Ok();
+            return StatusCode(200, company);
 
         }
